Implements the Encoder class.
The Encoder is responsible for composing the functionality of individual encoding stages, and exposes and interface that will be integrated in the top level interfaces for one-to-many and many-to-one cases.

This change also makes the strings used in example implementations unique, in order to avoid letting broken tests accidentally pass.

PiperOrigin-RevId: 225517306