Modifies the style of example implementations of encoding stages.
In particular, this avoids using repeated strings to refer to fields in params and encoded tensors dictionaries, and instead uses class level constants. This will make implementation of adaptive encoding stages simpler to understand, which typically wrap an existing encoding stage, changing how the parameters are controlled, but not changing the core encode decode methods.

PiperOrigin-RevId: 230003862