Preserve the order of variable lists and provide ModelWeights.keras_weights, and add methods to convert (keras model weights) <--> (tff model state).
This lets us get back from a ModelWeights structure to a list that can be used with keras_model.set_weights().

Fixes a bug where anonymous_tuple.flatten was used together with nest.pack_sequence_as, which happened to work before but didn't after the other changes in this CL. Consolidates all the TFF-specific code in optimizer_utils.py to a single function.

Raise a ValueError if someone tries to construct an AnonymousTuple with duplicated names, and also check to this when construct OrderedDicts of variables.

PiperOrigin-RevId: 234265469