Implements handling of tuples of length 1 in federated_zip.
In particular, federated_zip will now simply promote the federation of a 1-tuple of federated values, returning a federated 1-tuple of values.

WARNING: This CL is backwards-incompatible.
Now an IterativeProcess built through `build_federated_optimizer_process` will always return a *tuple* of metrics as the second return value of its next method, as opposed to returning a tuple if there are 2 or more metrics specified and a scalar if there is only 1.
PiperOrigin-RevId: 236763176