Replace assertCountEqual with assertEqual
assertCountEqual results in an equality comparison between different
elements. This results in an equality comparison between a string and a
Tensor. Without tensor equality enabled, this is a simple id check. With
Tensor equality enabled, this trying to convert the string to a Tensor
and compare Tensors by value. To avoid this, we simply use assertEqual
since it is sufficient for this test.

PiperOrigin-RevId: 262930422