Changes emnist.get_infinite() so that it generates consistent output from one use to the next.
The hash that's used to seed the random number generation is now a stable hash function (md5) via hashlib, replacing the Python built-in hash().  The built-in hash() has (as of Python version 3.3) internally defined randomness that causes inconsistent hash outputs for a given hash input when run in different processes. This was causing different infinite EMNIST images to be generated from one run to the next.

Also, fixes a few typos in the docstring for emnist.get_synthetic() method.

PiperOrigin-RevId: 263355721