Implements build_encoded_mean and build_encoded_sum.
This enables the tools from tfmot(.)tensor_encoding package to be used to specify encoding mechanism in the StatefulAggregateFn. The enclosed next_fn takes a state at tff.SERVER, derives parameters to be broadcasted to tff.CLIENTS, who encode their values, and only the encoded representation is sent to tff.SERVER. The decoding part is automatically split based on (potentially partial) commutativity with sum.

PiperOrigin-RevId: 266433175