Moves hash function and equality of anonymous tuple to use newly added name_array.
Previously, equality was checked by equality of name_to_index dict (IE equality as dicts, as unordered sets of key-value pairs), while hash was generated by iterating over the dict (where order matters).

Since dicts are unordered, it is an implementation detail of the underlying Python whether equality as AnonymousTuples then implies equality of hashes. This is not quite in line with the python spec of hash, which states that objects which compare as equal *must* return the same hash value.

PiperOrigin-RevId: 267644204