Fix static analysis, make this logic a bit more readable.
The type analysis did not know that unbound_reference_name and unbound_reference_type would always be None or not None at the same time. Combining the conditions removes the analysis error complaining about if unbound_reference_type is None it would not have a member attribute. Also this makes the logic a bit more clear.

PiperOrigin-RevId: 275083963