Add an example model using Keras + feature columns and a lookup table.
The unittest adding the example to FedAvg tests currently does not run because
TFF doesn't have its hands on the lookup table initializer (see
#783). This test is disabled by
default (since it won't pass) and any fix should enable and ensure it passes.

Switch to the local executor for tests to make them faster, and decrease the
size to 'medium' to reflect this improvement.

PiperOrigin-RevId: 280083440