Correctly compute the number of examples when Keras is using multiple inputs.
Previously tried calling shape on a nested structure of tensors, which will not
work. Instead flatten the structre and call shape on the first tensor (all
input tensors are expected to have the same first dimension size, the batch
size).

Updated example models and tests, ensured the test failed under the current
code and passes after this change.

Fixes #785

PiperOrigin-RevId: 280221692