Invalidate the entire cache when the underlying executor raises an execption.
This is a heavy hammer. We should follow-up, there maybe a better approach such
as subclassing RetryableError into an error that signals the underlying
Executor was restarted.

This avoids double delete errors in caching executor dictionary (a current
problem).  This can happen if two callers start awaiting a future from the
inner executor that fails and they both try to cleanup.

PiperOrigin-RevId: 284568338