Disallow empty datasets to be used for any type in the reference executor.
An empty dataset will return an empty list which will pass type spec checking,
regardless of the actual type of the dataset elements. Instead examine the type spec
of the dataset and compare to the requested type.

This is in preparation to make the "local executor" stack the default execution mode.

PiperOrigin-RevId: 288401901