Delete cache across executor close
The test added here fails before this CL with "futures are tied to different event loops"

This CL clears the cache in the caching executor on close.
Other executors, such as the concurrent and remote executors,
rely on values they produce not being reused across a call to `close`.

It may be that we need to change when `close` is called in order to allow
some resources to be reused between top-level calls to federated computations,
but in any case we need for the cache to be cleared at the same time that
the remote executor or concurrent executor threads are cleaned up.

PiperOrigin-RevId: 291426443