Implement Dispose and call it from RemoteExecutor
This CL attempts to address b/139137514 by disposing
of remote worker values whenever the corresponsing
local `RemoteValue` is no longer in use. It does this
by attaching to the weakref finalizer method so as not
to force all clients (executors and executor users)
to manually call a destructor on every ExecutorValue,
which would be very intrusive and likely error-prone.

However, this solution comes at a somewhat significant
cost, in that there's no guarantee when destructors
will run-- they're tied to garbage collection. This
also makes it difficult or impossible to write a
reliable test for them.

One alternative solution would be to dispose of all
values associated with a RemoteExecutor when `close`
was called on the RemoteExecutor. However, previous
issues (see cl/291464360 and b/148288711) have
indicated that sometimes values are being used after
their associated executor has had `close()` called
on it, making this solution potentially bug-prone.
Additionally, this wouldn't allow us to clean up
resources until after a call to `close()`, which
is potentially undesirable given the ongoing
discussion pushing for `close()` to be called at
program end rather than between rounds.

Note: if the caching executor holds onto remote
executor values, they will continue to persist.

Testing: this change was not tested for correctness,
only checked for absence of new errors.
PiperOrigin-RevId: 292401078