Clarify the lambda executor
Previously, the lambda executor included a number of
hot-swaps between functional and non-functional types
that could result in reduced accuracy error messages.
This change removes that conversion in favor of
preserving the original types of values.

Additionally, `_evaluate` is changed to be called
eagerly-- `pb.Computation`s are no longer stored
except as the internals of a new `Lambda` class
which encapsulates a `pb.Computation` lambda and
its containing scope.

PiperOrigin-RevId: 294997379