This CL provides the utils needed for implementing a personalization evaluation experiment. Specifically, it provides a `build_personalize_fn` helper function, which builds a personalization strategy when called.
A personalization strategy is represented as a `tf.function` called `personalize_fn` (see the doc in `p13n_utils.py` for more details). The file also contains a `tf.function` called `evaluate_fn` which evaluates a model on the given dataset.

This is Part I of a larger CL that gives a personalization experiment on EMNIST-62 dataset. Part II will provide a training loop using the utils implemented in this CL.

PiperOrigin-RevId: 297407622