Remove the ability to call `tff.framework.set_default_executor` and `tff.framework.set_default_context` with `None`.
This functionality allowed users of TFF to throw away the execution context. This released some objects from memory, but should not be needed. The related memory issues should be fixed and this API is overloaded in a confusing way.

Additionally, this change is a precursor to inverting the `context_stack` -> `executors` dependency, which should make testing the `executors` package much easier.

PiperOrigin-RevId: 297831094