Extract the logic to compute a federated broadcast from `FederatingExecutor` and `ComposingExecutor` to the `executor_utils` module.
* Added `compute_intrinsic_federated_broadcast` function to the `executor_utils` module.
* Added tests for `compute_intrinsic_federated_broadcast` function.
* Updated the `create_dummy_federated_type_*` functions to be more naturally named.
* Updated `create_dummy_federated_type_at_clients` to be parameterized by number of clients.

Reverted a change to use the same instance of a child executor in a federating executor stack. A new instance of each executor is constructed instead of using references to the same instance of an executor. While not required, this ensures that the state on each executor does not change the behavior of these tests.

PiperOrigin-RevId: 308276876