Remove unnecessary functionality from `FederatingExecutor` and `ComposingExecutor`.
This change attempts to align how `FederatingExecutor` and `ComposingExecutor` implement the `executor_base.Executor` API. In many cases these implementations are functionally the same but they are implemented differently. As a first step to reducing this duplication, this change updates the implementations to actually be identical. This makes reducing the duplication in a follow up easier.

There is some unnecessary functionality. In some cases, these executors contain functionality which makes it easy to write hard to read tests. Therefore this change is also a step towards improving our test culture.

* Removed handling `computation_types.NamedTupleType` in `create_value`.
* Removed handling `pb.Computation`s of types: placement, call, tuple, and selection in `create_value`.

PiperOrigin-RevId: 313625601