Fix visibility involving placements in the `impl` package.
Note: That this was a simple regex search-replace, formatting, and BUILD file edits; no logical changes.

Essentially the `placements` package is forward declaring the `placement_literals` package. This change helps to clean up some of the circular dependencies (Bazel) and potential circular imports (Python) that are caused by the `impl` package depending on the `api` package.

In some cases we were actually importing both `placements` and `placement_literals`, which simply causes the implementations to be more confusing.

PiperOrigin-RevId: 313840517