Uses parameters of incoming intrinsics rather than arguments to construct parameter to merged intrinsic.
We were previously using the types of arguments to the merged calls to populate the parameter type in the merged intrinsic. This is not completely general, as the test from which we remove the skipTest illustrates.

In particular, we will always call the intrinsic on a type which is assignable to its parameter, but this does not imply that these types are equivalent.

This can burn us in the case, for example, that we have an intrinsic declaring a tensor parameter with undefined dimensions, called on an argument with tensor of concrete dimensions. The previous implementation would attempt to construct a merged intrinsic which only accepts parameters of tensors with this exact concretized dimension, rather than read the parameter type from the function parameter declaration.

PiperOrigin-RevId: 314151869