Call `tff.federated_zip` on the output `build_clip_norm_aggregate_fn.next` to promote the placement to the "top" of the type.
For example, if we have a tuple of placed values `(U@S, V@S)` this will promote
the placement outside the tuple `(U, V)@S`. Values that already have placement
at the root are unchanged.

This is in preparation of extending `tff.learning` to accept
`tff.templates.MeasuredProcess` instead of `StatefulFn`.

PiperOrigin-RevId: 316785026