Fix error-checking logic for `get_canonical_form_for_iterative_process()` for `IterativeProcess`'s with no aggregation.
The logic was previously failing for this case, but with a harder to grok error message.

The previous logic had a slight bug in the branching logic, such that we would never hit the explicit `ValueError` in the final `else:` block.

The tests were also incorrectly passing because they hit a different validation check:

```
ValueError: Expected an AST containing an intrinsic with the uri: federated_secure_sum, found none.
```
PiperOrigin-RevId: 322385240