Implements `resolve_higher_order_functions`.
This change introduces a transformation which guarantees that any instances of a `Call` building block in its return value are either to a concrete function, or a call to a function accepted as a parameter to an uncalled function.

By setting up this invariant, complexity in other transformations can be significantly reduced.

PiperOrigin-RevId: 322402665