Replace default_executor with set_local_execution_context
Previously, the default context included a local execution
stack, but did not include the compiler built to optimize
for this native backend. This meant that users of TFF were
not getting the advantages of compilation-for-native unless
they explicitly set `set_local_execution_context`.

Additionally, this change moves the integration tests inside
`core` to use this same local execution stack, including the
native compiler. These tests are designed to exercise the
local execution stack, and shouldn't be testing against a
different environment than the one users experience.

PiperOrigin-RevId: 324705661