Fix Shakespeare preprocess bug where the last character in `CHAR_VOCAB` maps to the same id as OOV.
The underlying reason behind this bug is due to the way that OOV is handled in `tf.lookup.StaticVocabularyTable`. The table uses the number of possible values to determine the OOV id. This means that if the keys and values in `tf.lookup.KeyValueTensorInitializer` are initialized at an offset (like it was here previously), there will be collisions with OOV.

e.g.
```
lookup_table = tf.lookup.StaticVocabularyTable(
    tf.lookup.KeyValueTensorInitializer(
        tf.constant(['a', 'b']), tf.constant([2, 3], dtype=tf.int64)),
    num_oov_buckets=1)
lookup_table.lookup(tf.constant('a'))  # 2
lookup_table.lookup(tf.constant('b')) # 3
lookup_table.lookup(tf.constant('c'))  # 2
```
PiperOrigin-RevId: 326599875