Add a `shuffle_clients_before_split` flag.
This flag decides whether to shuffle the clients before splitting them into training and p13n sets.

This cl also prints the mean accuracy of the p13n model obtained at Epoch 1. The goal of this change is to highlight that one can access the metrics for each epoch obtained by p13n.

PiperOrigin-RevId: 327130782