Move `ReferenceExecutor` to the `backends` package.
The impl package was quickly becoming unmaintainable and hard to reason about. In an effort to create order, modules are begin moved into sub packages with stronger visibility rules.

In this case specifically, we have created a backends package that contains other contexts that act as backends to TFF, so it was a natural destination.

* Create `backends/reference` package.
* Moved `ReferenceExecutor` to `ReferenceContext`.
* Added `set_reference_context` function.
* Removed `tff.test.ReferenceExecutor` API.
* Updated callers.

PiperOrigin-RevId: 328167081