Add `golden` library for writing golden tests
We have a number of tests against the string
format of large ASTs that are painful to update.
This library should allow us to move them out
into separate files and update them easily.

Similarly, we'd like to introduce assertions
against of user-facing error messages in some
places to ensure that they're readable and
high-quality, but string assertions within
Python are annoying and difficult to update.
Assertions against an automatically-updateable
golden file should make this much easier.

PiperOrigin-RevId: 332963417