Improve formatting of type mismatch errors
Previously, these would always show the stringified
version of the type, which delegated to the compact
representation. Aside from causing confusing line
wrapping, this representation didn't display
container type mismatches or other things that are
only present in the `repr`, so it could wind up
reporting two not equal types with the same string
representation. This fixes that issue, as well as
providing some other niceties like singleline vs.
multiline based on message length and diffs for
large type mismatches.

PiperOrigin-RevId: 335498789