Add BroadcastForm to represent new task type
This is a sibling
to CanonicalForm which is intended to represent a
new task type which involves only server work +
broadcast + client work with no final aggregation.

My hope is to reuse much of the existing CanonicalForm
compilation infrastructure for compiling to this form
(that is, sharing much of canonical_form_utils.py
between both forms). An alternative approach would be
to slightly extend and re-use CanonicalForm by adding
a client output and checking to ensure that the arguments
to the aggregation functions are empty structures. I'm
not strongly opposed to that formulation, but have
developed a weak preference for this approach in order
to reduce overloading and layering of complexity on top
of the existing system. My hope is that the approach taken
in this CL will encourage work towards easier
development of varied computation "shape" definitions.

PiperOrigin-RevId: 337184931