Fixes issue with `float('inf')` not being `np.inf`.
`float('inf') is np.inf` evaluates as `False`, but `float('inf') == np.inf` and `float('inf') in [np.inf]` evaluates as `True`, which leads to empty AssertionError if `float('inf')` is provided as we surprisingly do not catch it earlier.

Changing this to use `np.isinf()` to support different ways to express infinity.

Also extends the tests to build `AggregationProces`, some of which would have failed.

PiperOrigin-RevId: 344475293