Use a configurable default for num_clients
Previously, we would allow local inference of the number of clients
to run on remote workers to fail, falling back to num_clients
configured on the workers themselves. This was fragile, split
configuration across multiple places, and caused confusing error
messages.

This change instead uses a default num_clients which is overrideable by
users, allowing computations in which cardinality cannot be inferred to
coexist with ones where there are different numbers of clients.

PiperOrigin-RevId: 356859451