Flatten selections to index in computation.proto
Previously, selections of the same index could happen
via either string name or numeric index. This required
additional branching through significant portions of the
codebase, but at little additional value: the name of the
selection can be recovered for free from the type of the
source.

This is currently written as a backwards-incompatible change
to computation.proto and executor.proto. As far as this author
was able to gather, there is no current persistence of values
of either proto, so this should not cause issues. If this
proves to be incorrect, this CL should be easy to roll back
and re-attempt via a more gradual transition.

PiperOrigin-RevId: 359951721