Update test sizes.
Updated test sizes according to the following logic:

* Small tests run in a single process and must not block.
* Medium tests run across multiple processes on the same machine and may block.
* Large tests run across multiple machines.

In this context, "block" means "wait on a system or event outside of the test process". This includes reading from a disk, accessing a network, or sleeping for a fixed period of time. Small tests may still use blocking concurrency internally so long as at least one thread in the test process is making progress at all times.

PiperOrigin-RevId: 364333175