Remove unused federated_reduce intrinsic
There was a not-insignificant amount of code involved
in providing this intrinsic. However, it isn't being
used at all, in particular because it isn't possible
to use in production (where multi-layered aggregation
is a requirement). Instances of federated_reduce that
use an associative binary accumulate operator can be
easily transitioned to federated_aggregate, as was
already done in the _federated_reduce_with_func helper
in federated_aggregations.py.

PiperOrigin-RevId: 365091411