Unify _process documentation and special-case next_fn argument errors
Previously, a mismatched state type in a _process would result in a
long error message showing the full return type of `initialize_fn`
and the full parameter type of `next_fn`. This is necessary because
the state may be the first element of `next_fn`'s parameter or the
*whole* of  `next_fn`'s parameter. However, for `AggregationProcess`
we can do better since we know that it *must* be the first element.
This CL adds an optional argument to signal whether or not the state
type is all or part of `next_fn`'s parameter, and uses it from
`AggregationProcess`.

PiperOrigin-RevId: 365138435