Force close the function wrapper generators when they are done being used to trigger `finally` blocks that pop the context from the stack.
It seems like the generator may be hanging around in some situations without
closing, causing the context stack to not be popped when finished. This seems
possibly related to:
https://docs.python.org/3/reference/expressions.html#yield-expressions

"Yield expressions are allowed anywhere in a try construct. If the generator is
not resumed before it is finalized (by reaching a zero reference count or by
being garbage collected), the generator-iterator’s close() method will be
called, allowing any pending finally clauses to execute."

PiperOrigin-RevId: 365932207