Add `tf_computation` I/O support for `tf.{RaggedTensor, SparseTensor}`
Before this CL it was not possible to use a `tf.RaggedTensor` or `tf.SparseTensor` as an input or output of a `tf_computation`. This CL unlocks that possibility, though it does not yet provide users with an easy way to spell the type of such an
operation.

After this CL, these values are stored in their dense representations using the existing support for Python structures. In order to prevent accidental summation of the dense representation of RaggedTensor and SparseTensor, `federated_sum`'s checks are extended to ensure that no `RaggedTensor` or `SparseTensor` container types are present. However, this check does not extend without effort to other non-intrinsic
aggregations, so ensuring that the dense representations of RaggedTensor and SparseTensor are not naiively summed is left to library authors and end-users.

PiperOrigin-RevId: 367261866