Statically intern TFF types for the life of the Python runtime
This change removes the WeakValueDictionary which was causing weird behavior
where lookups would occur after parts of the Python standard library had
already destroyed themselves. `atexit` hooks to clear the dictionary appeared
to fix this at first, but it regressed again (for unknown reasons, this time
on a smaller subset of programs).

TFF types themselves are not particularly large in memory, and are frequently
long-lived anyways (many are in global-scope objects), so the longer lifetimes
are likely not a significant downside.

PiperOrigin-RevId: 373052202