Add a tutorial that demonstrates how TFF can be used to train a very large model where each client device only downloads and updates a small slice of the model using`tff.federated_select` and sparse aggregation.
Concretely, we consider logistic regression for multi-label classification, predicting which "tags" are associated with a text string based on a  bag-of-words feature representation. Importantly, communication and client-side computation costs are controlled by a fixed constant (`MAX_TOKENS_SELECTED_PER_CLIENT`), and *do not* scale with the overall vocabulary size, which could be extremely large in practical settings.

PiperOrigin-RevId: 376043544