Update tensorflow_federated xla backend for changes to the XLA Python client.
In a future change, I plan to remove the `xla_client.get_local_backend()` API, which chooses a "good" backend and caches it. The code duplicates logic in JAX which is its primary user.

Instead, xla_client will only expose factory methods for backends; any caching or heuristic choice of backends is up to the user.

Here, use the JAX version of the API instead.

PiperOrigin-RevId: 376842775