Implements a lower level composition of a `LearningProcess`.
This change defines `DistributionProcess`, `ClientWorkProcess`, `FinalizerProcess` as specialized `MeasuredProcess` instances that together with existing `tff.templates.AggregationProcess` form a general structure of a `LearningProcess`.

Implements a `compose_learning_process` method, that given these processes, checks their compatibility, and returns a composed `LearningProcess`.

These symbols are not added to the API at this point. For now, they are intended for prototyping and some details are likely to change.

PiperOrigin-RevId: 379881518