Enables progress with partially available workers in CPP.
The bug in the previous implementation: since we currently create channels on the fly during stack construction, we were stuck in a loop of failing to give these channels appropriate time to attempt connection. They were constructed and destroyed too to ever either enter a failure state or achieve CHANNEL_READY status.

This change introduces a duration parameter and allows the filtering function to wait up to this time on all channels to be ready, filtering to tasks which register as connected after this deadline.

PiperOrigin-RevId: 385251570