Fixes an issue with specs provided to optimizer in model delta client work.
Previously, the tensor specs provided to the optimizer were for the entire model weights, while the expectation is that it shuold manipulate only trainable model weights.

This also restructures the code to make sure optimizer does not work with lists but tuples instead, as if the optimizer state contains list of tensors, the "for _ in data:" pattern will try to stack all of the tensors, raising error if the tensors to not have the same rank - e.g. when they include bias and kernel of a layer being optimzied.

PiperOrigin-RevId: 387019484