Replace value indices in MapReduceForm with descriptive names
These value indices were fragile and would all require replacement
upon the addition of new intrinsics to MapReduceForm. This change
provides more descriptive names and stablilizes the selection of
values to be resilient to new intrinsic additions.

This change is designed to make it eaiser to add MapReduceForm
support for federated_secure_sum and federated_secure_modular_sum.

PiperOrigin-RevId: 387458471