Returns second argument to binary operator in create_binary_operator_with_upcast uncoerced in the case of type assignability to the first arg.
This change represents a slight relaxation in (static) semantics for create_binary_operator_with_upcast. Previously a binary operator with signature like:
```
<float32[?],float32[1]>
```
could not be constructed, though TensorFlow's shape assignability allows for a graph like this. This change allows for signatures such as the above to be represented in TF computation factory.

We do not, however, open the door to signatures like:
```
<float32[?],float32[1, 1]>
```
(though TF will generate graphs which represent it), since such a signature does not pass TFF's notion of assignability.

PiperOrigin-RevId: 390188849