Fix a bug in cache_key generation for no-arg computations that return no tensors.
Include the type signature of the computation when hashing the computation and
assigning IDs. Without the type signature the following two computations would
be the same:

```
(<> -> <> )
(<> -> <<>>)
```

This is because the TensorFlow graph for both only includes a single `NoOp`
node, resulting in the same hash. However, executor should return different
Value protos for these two computations, and not treat them as identical.

PiperOrigin-RevId: 399195391