Implements tff.structure.Struct interface for structure-typed instances of tff.Value.
This change allows for the use of tff.structure.map_structure on structures of values inside of federated computation bodies, but does not go as far as it might.

In particular, this change preserves the invariant that every Python value in the body of a federated computation is a TFF Value, and hence does not propagate user-defined containers into those bodies. This change leaves a code comment indicating how this can be done if desired.

Further, this change does not handle federated structures, just structures of (possibly federated) values.

PiperOrigin-RevId: 401529505