Remove `federated_collect`
`federated_collect`, which could be used to transform a client-placed value into a
server-placed sequence of values, is currently unused in Google and cannot be used with MapReduceForm. By materializing values from client devices directly onto the server it presents a significant privacy hazard.

Additionally, `federated_collect` has been reported to cause confusion in researchers and other users of the TFF native runtime. These users have reported that they reach for `federated_collect` in order to materialize a value back into Python, but this transformation is unnecessary: client-placed values may be returned directly as outputs of computations running on the native runtime.

PiperOrigin-RevId: 401565220