Test dataset reduction computation in tensorflow_executor_test
As an aside, this CL adds a new `main` function for tests running in OSS.
This was necessary in order to use absl flags in OSS C++ tests because OSS
gtest_main intercepts argc and argv but does not parse absl flags. For
details, see https://github.com/google/googletest/issues/3646.

PiperOrigin-RevId: 407875061