Remove the CachingExecutor
The CachingExecutor formerly played a critical role in the TFF execution stack, as it prevented multiple execution of the myriad duplicate computations in TFF's ASTs. Nowadays, TFF no longer generates these types of duplicate computations, and it is no longer to cache at this level to ensure performance.  Additionally, this type of caching plays poorly with nondeterministic computations.

Prior to this CL, the CachingExecutor was already unused outside of tests. This should be a no-op for TFF customers.

PiperOrigin-RevId: 411112532