Deletes reference context.
With migration of form_utils_test, the reference context / executor will be dead code.

Further, there are outstanding bugs indicating that the reference context should be removed from the public API, and used in unit tests--however, it is not used for any unit tests any longer. Therefore, preferring to remove.

PiperOrigin-RevId: 413804412