Updates async execution contexts to be usable with ConcreteComputation's __call__.
Introduces a cardinalities lock to ensure cardinalities are not swapped out on a worker in the middle of an invocation. This is relatively overkill; in a local computation invocation, this would not be necessary, which means we in fact could imagine a world in which this would not be necessary at all (workers hosting multiple versions of the runtime in parallel rather than one). Still, I think the benefit here is unclear for now, so leaving as-is with the locking.

Ensuring a instance-consistent cardinalities while preserving maximum parallelism is nontrivial; this CL introduces a shared class to manage. If we relaxed this requirement (and supported multiple concurrent cardinalities) we could remove this complexity. Happy to consider other approaches here.

PiperOrigin-RevId: 415562949