Moves fed_avg `client_work` component from learning/algorithms to learning/templates.
The generality comes in support for weighting by `tff.learning.ClientWeighting`, and in support for both Keras and TFF optimizers. This also makes it available in the API.

This change also fixes a bug in the client_work component which did not work with a stateful TFF optimizer, due to strange interactions of `tf.data` and autograph.

PiperOrigin-RevId: 420121293