Update `tff.program.CSVFileReleaseManager` to make the fieldname of the key configurable.
* Updated the `tff.program.CSVFileReleaseManager` constructor with a default value.
* Updated all existing call sites to retain existing functionality.

This change adds the capability to customize the output of the `tff.program.CSVFileReleaseManager` for a specific context. The default value reflects the recent API change, the `round_num` parameter from `tff.simulation.CSVMetricsManager` was renamed to `key` in `tff.program.CSVFileReleaseManager` and represents an integer use to reference the released `value`. It often maps to round number, but it doesn't have to map to round number.

PiperOrigin-RevId: 421382222