Implement a `secure_sum_then_finalize` metrics aggregator.
This aggregator only supports fixed quantization ranges because it is intended
to be used with `tff.learning.build_federated_evaluation` which is stateless.
Internally it uses the SecureSumFactory which uses 32 bits for communication,
and provide the default range of [0, 100_000.0] this has shown to work well
empirically for typical evaluation metric values.

PiperOrigin-RevId: 422685925