Compile computations to TensorFlow via traced evaluation
Previously, computations were compiled to TensorFlow using a syntax-directed translation approach in which subcomputations were each transformed to TensorFlow, and the resulting subgraphs were merged upwards to produce the final TensorFlow graph. This approach had two significant downsides.

First, it was complicated: generating TensorFlow for all varieties of subcomputations required a complicated series of pattern-matching, shaping, and reshaping, as well as manual construction of the appropriate TensorFlow graphs.

Second, merging these subcomputations was very slow: see b/201294306#comment14 and the thread above for discussion of the significant time spent merging graphs. The new approach to TensorFlow generation is multiple orders of magnitude faster. See bug link for timing and examples, with one large workload moving from 412s in translation to only 9s.

N.B.: This CL introduces some visibility patches that are necessary in order to take a dependency on the TensorFlow context from the TensorFlow transformation code. The code reuse this enables is significant, and it does not create a target dependency cycle. After this change lands, much of the existing transformations code can be removed, after which the remaining TensorFlow-specific transformations can be moved to a new subpackage living above the TensorFlow context.

PiperOrigin-RevId: 423178193